Now getting latest in bs

+++ src/endpoint/win/downloads_msi.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-EPVER=(cat latest.text)
+[ -f "latest.txt" ] || exit 1
+EPVER=$(cat latest.txt)
 MSIFN=mydlp_$(echo $EPVER|sed -s 's/\./_/g').msi
 echo "Generating latest.msi for $EPVER"
 [ -f "latest.msi" ] && mv -f latest.msi $MSIFN
